requirements.txt
aiogram~=3.2.0
aiogram_dialog
# opencv-python~=4.8.1.78
requests~=2.31.0
python-dotenv~=1.0.0
aiohttp~=3.9.1
numpy~=1.26.2
pytz~=2023.3.post1
pydantic~=2.5.2
cachetools~=5.3.2
coloredlogs~=15.0.1
alembic==1.13.0
asyncpg==0.29.0
SQLAlchemy~=2.0.23
redis